Have workers process multiple bundles
Initially, each mapper received one bundle, figured out the chunks in
that bundle, and distributed each chunk to a worker. That makes sense
for queries with a relatively small number of large bundles. In
contracty, the HCA will at first have bundles with data from a single
cell, so we should expect queries with a large number of very small
bundles.

This change adjusts the driver, mapper, and worker so that multiple
bundles are distributed to a single worker. This reduced the number of
lambdas and the number of concurrent writes to the output zarr store.