parallelize read from DSS in workers (#94)
* parallelize read from DSS in workers

* Use thread-safer DCP interface

This incorporates the change in HumanCellAtlas/dcp-cli#202

* Include thread id in logs

This is helpful in our multithreaded code paths

* Don't print logs twice

AWS Lambda has a default root logger that captures everything that goes
to stdout. If you don't turn `propagate` off, you get every message
twice in cloudwatch.

* Poll for locks much less frequently

A failure mode we encounter is using up capacity on the lock table. If
we're anticipating smaller numbers of worker lambdas that handle larger
blocks of works, we can wait longer between checking for lock
availability. This reduced our dynamo consumed capacity.

* Distribute 100 bundles to each lambda

And aggressively parallelize DSS I/O.

* Assume that each bundle has a single cell in the mapper

This is a bad assumption, but it's helpful for the demo.

* Address comments