Change from donor to specimen
One property of donor was "organ", so one row in the donor table would
have a donorkey and an organ. It turns out that one donor (so far)
contributed two different organs, which were processed separately. There
wasn't any way to capture that in the donor table, so this switches
donor to specimen.