Create endpoint to subscribe to DSS notifications (#217)
* create DSS notifications handler API endpoint

* implement DSS notifications lambda handler

* refactor Redshift loader script to use notifications logic

* Set diststyle and sorting for analysis and feature tables

* Collapse expression counts over identical gene/transcript ids
This can happen if there are, for example, entries for ENSGxxx and
ENSGxxx.PAR_Y. We don't want duplicate entries making it to redshift.

* Calculate 10x cellkey using bundle id
The previous method doesn't produce unique cell ids for a number of
reasons.

* Handle cell lines when getting organ info

* Change from donor to specimen
One property of donor was 'organ', so one row in the donor table would
have a donorkey and an organ. It turns out that one donor (so far)
contributed two different organs, which were processed separately. There
wasn't any way to capture that in the donor table, so this switches
donor to specimen.

* update loader query to include SS2 and 10X bundles

* add unit tests for ETL and transformers, fix lint

* fix scipy version

* Extend s3 preload bucket expiration to 30 days

* replace Redshift run_query with transaction

* refactor Notifications lambda to Notification

* vendor packages for chalice deployment

* deploy dev locally