Reduce memory consumption in conversion job (#267)
Update the conversion jobs to use a small amount of memory that (mostly) doesn't depend on the request size. Previously, lots of data was being loaded into memory at once, which caused requests for ~1 million cells to use a huge amount of memory. This restructures the conversion so data from only a few cells is in memory at any given time. The tradeoff is that this is pretty slow.

Also, this adds a much more extensive functional test of the conversion jobs.